1. General code cleanup and minor fixes 2. Since I got a gecko now changed some of the debug statements to use the gecko and removed my other way of debugging 3. HTTP Download would sometimes core dump if the file length was 0 4. Hopefully fixed video overscan problem when coredumps occured. NEEDS more testing with Widescreen and normal screen mode. 5. When NusGetObject was given a version 0 it was saving the folder as v0. It now checks for a TMD and sets the version based on the downloaded TMD to save the files into the correct folders. Will need to revisit this some more to help support offline mode better when passing a version 0 to the downloader. 6. Shop Channel was hardcoded to v18. It will now download latest version 7. Installing some of the channels were not region specific. Example Korea has a different shop channel then the other Wiis.

+++ source/patchmii_core.h
@@ -6,7 +6,7 @@
 void patchmii_network_init(void);
 // Call with version = 0 to get the latest version
 // Set "patch" if you want to try to patch title contents
-s32 patchmii_install(u32 in_title_h, u32 in_title_l, u32 in_version, u32 out_title_h, u32 out_title_l, u32 out_version, bool patch, bool patch2);
+s32 patchmii_install(u32 in_title_h, u32 in_title_l, u16 in_version, u32 out_title_h, u32 out_title_l, u16 out_version, bool patch,bool patch2);
 s32 install_temporary_ios(u32 base_ios, u32 base_ver);
 s32 load_temporary_ios(void);
 s32 cleanup_temporary_ios(void);
